// angular-integrations/code-grant-oidc-single-app-integration/projects/private-app/src/app/rest-client/model/transactionDetail.ts
/**
 * Swagger Customer - OpenAPI 3.0
 *
 * Contact: apiteam@swagger.io
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { PaymentMethodEnum } from './paymentMethodEnum';
import { DomainEnum } from './domainEnum';
import { TransactionStatusEnum } from './transactionStatusEnum';


export interface TransactionDetail { 
    id?: string;
    transactionNumber?: string;
    sequenceNumber?: number;
    date?: string;
    domain?: DomainEnum;
    location?: string;
    value?: number;
    status?: TransactionStatusEnum;
    paymentMethod?: PaymentMethodEnum;
    taxAmount?: number;
    netValue?: number;
}
export namespace TransactionDetail {
}