// angular-integrations/code-grant-oidc-single-app-integration/projects/private-app/src/app/rest-client/model/sortOrderEnum.ts
/**
 * Swagger Customer - OpenAPI 3.0
 *
 * Contact: apiteam@swagger.io
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export const SortOrderEnum = {
    Asc: 'ASC',
    Desc: 'DESC'
} as const;
export type SortOrderEnum = typeof SortOrderEnum[keyof typeof SortOrderEnum];