// angular-integrations/code-grant-oidc-single-app-integration/projects/private-app/src/app/rest-client/model/transactionFilterResponse.ts
/**
 * Swagger Customer - OpenAPI 3.0
 *
 * Contact: apiteam@swagger.io
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { TransactionDetail } from './transactionDetail';


export interface TransactionFilterResponse { 
    data?: Array<TransactionDetail>;
    foundItems?: number;
    totalItems?: number;
    previousPageToken?: number;
    nextPageToken?: number;
}